requirements.txt
pygame
PyOpenGL
numpy
sounddevice